inputDataLatihan
