+++ lib/views/inputdata.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class InputData extends StatefulWidget {
+  const InputData({Key? key}) : super(key: key);
+
+  @override
+  _InputDataState createState() => _InputDataState();
+}
+
+class _InputDataState extends State<InputData> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+      title: Text('Input Data Page'),
+    ));
+  }
+}
